Fix issues evolving an extension's models.
We weren't properly calling Django Evolution's evolve management
command. We needed to go through execute() and not evolve(), in order to
get objects like stdout and stderr set up. Without this, we get an
exception when Django Evolution tries to output any text.

While fixing this, I changed it to take the output from the evolve
command and include it in the stored error. It's now shown to the user,
and logged, when it fails. It also logs when it succeeds and there's
something to show.