Fix building translations on pristine systems
If we are attempting to build on a system without an existing copy of Djblets
installed in a reachable location (such as distribution packaging), we need to
ensure that the source directory is in the PYTHONPATH or the import of
djblets.util.filesystem will fail.

Testing Done:
With this patch, I was successfully able to build Djblets for Fedora.

Reviewed at https://reviews.reviewboard.org/r/5300/