Make dashboard column selector sort on detailed_label.
The dashboard column selector was sorting based on the label field which is
empty for many columns and different from the labels shown in the selector for
many others. The result of this was that the list of available columns showed
up in a seemingly random order.

Given that some columns (like the check-box column) include HTML in their
detailed_label, I've split this into two different parameters to __init__:
detailed_label, and detailed_label_html. We sort based on the first, and
display the second.

This requires some changes in Review Board to have everything pass in the data
to __init__ correctly.

Testing done:
Looked at the column selector and saw that everything was in alphabetical
order, even when the column included HTML or other entities at the beginning of
the text.

Fixes bug 3260.

Reviewed at https://reviews.reviewboard.org/r/5537/