Fix static media building issues with the config forms.
The config forms were referencing admin media and other Review Board
assets, which worked when in the Review Board tree, but didn't work in
Djblets.

For now, we're just going to bundle the icons we need. We'll want to
revisit this at some point to use spritesheets and retina images.