Add some presentation improvements for configforms.
This adds a new class used for the descriptions at the top of a config
form, used to provide decent spacing when using one or more paragraphs.

It also shows a spinner when the item indicates it's communicating with
the server, and hides it when it's finished communicating.

Reviewed at https://reviews.reviewboard.org/r/5696/