Fix some rendering-related problems with ListView and ListItemView.
ListView was attaching to collection events in initialize and rendering
as a result. This meant that they could render before render() was
called, causing breakages. Now, these events are attached in render()
instead.

ListItemView had some whitespace before the default displayed text,
which isn't ideal. This changes the template to not indent.