Add a LocalDataQuerySet for using generic data in WebAPIResource.
This introduces LocalDataQuerySet, which operates similarly to a
QuerySet, but can store any iterable data. The interface it exposes is
similar enough to QuerySet for basic usage.

While this may be useful for various purposes, it's largely intended for
use with WebAPIResource. Before now, getting WebAPIResource to return
data not from a database was tricky. You had to rewrite the
serialization functions, get(), get_list(), etc. Now, get_queryset() can
return a LocalDataQuerySet, passing in whatever it needs, and
WebAPIResource will work with it.

Reviewed at https://reviews.reviewboard.org/r/5880/