Cache object lookups in WebAPIResource.get_object.
It's pretty common for different parts of a WebAPIResource subclass to
perform a get_object() call either on itself or on another resource.
This will happen in get_queryset() and in the permission-checking
functions. These checks are time-consuming and result in excess
database queries.

The built-in get_object() now supports caching of these objects. The
caches are tied to the request, so any redundant get_object() calls
taking place within that request will just use the original call's
result.

Reviewed at https://reviews.reviewboard.org/r/5906/