Fix saving extension settings.
A recent change added a 'request' parameter to SiteSettingsForm, and
changed the calls in siteconfig to pass the 'data' and 'files'
values as keyword arguments.

This broke the form, because extensions use the configure_extension
view, which passes data as positional arguments.

While that view could also just be updated for this, the convention is
to pass 'data' and 'files' as positional arguments, and most callers do
so. To provide the best compatibility, we now do this as well, and leave
optional keyword arguments (request) as keyword arguments accessed
through `kwargs`.

Reviewed at https://reviews.reviewboard.org/r/5967/