Fix the extension 'enabled' state in the API after enabling/disabling.
A recent regression caused ExtensionResource's 'enabled' state to be
incorrect after enabling/disabling an extension. The value would be that
of the old state. This led to an unexpected combination of states in the
extension list, which caused the "Enable", "Disable", etc. actions to
disappear.

The reason this happened is that get_object() now caches results, and
ExtensionResource was attempting to use it in order to reload the
RegisteredExtension instance. We now reload it like all other resources
do, which is to just re-fetch from the database using the existing
instance's ID.

Reviewed at https://reviews.reviewboard.org/r/5969/