Improve style customization of the configforms templates.
This enhances the configforms in a few ways to allow for more flexible
styling.

First, the boxes now make use of a "box-foot" class, which is used on
RBCommons. This extensions our box style convention to have a "head",
"main", and "foot", allowing for a separation of CSS rules for the main
part of a box and the footer.

It's also now possible to specify the base template name through a
base_template_name attribute on the ConfigPagesView.

Field rows will now contain a 'checkbox-row' class if it's a checkbox,
radio button, or a multi-select of checkboxes.

Finally, there was a typo where "clas=" was used instead of "class=",
which has been fixed.

Reviewed at https://reviews.reviewboard.org/r/5995/