Fix config forms item action and spinner styles in Firefox.
Firefox didn't display the item actions very well. They were being
wrapped to the following line, due to a combination of `float: right`
and `white-space: nowrap` (ironically).

To fix this, we now put all the action elements in a new <span> that's
floated right and placed as the first element in the item. This gives us
the look we want across all browsers.

The spinner is also now properly vertically aligned, and placed right
before any actions (if there are any).

Reviewed at https://reviews.reviewboard.org/r/5996/