Avoid deleting previous PYTHONPATH
With the previous putenv command, if an user had already set the PYTHONPATH environment variable, its value was deleted and substituted by os.getcwd(), which may prevent the installation if the dependencies are installed in a different path. Now, it appends this value to the previous value set in the sys.path, so the user configuration is kept.