Fix a couple style issues in config-forms.less.
This changes one of the rules to use '.box-title' instead of 'h3',
for consistency with other rules.

It also changes the subsection styles to match what we had in RBCommons.
Nothing but RBCommons seems to use this style currently, and I don't
know why the style differed from what we had there, but worth moving it
back.

Reviewed at https://reviews.reviewboard.org/r/6004/