Make middle or control-clicking on dashboard cells work properly.
Dashboard cells with links had a long-standing bug where if you middle or
control-clicked on the link, it would open it in a new page, but if you middle
clicked on the row, it wouldn't. This was doubly confusing because a plain left
click on the cell would open the link as intended.

The issue here was that clicks inside the cell were handled via javascript.
After trying unsuccessfully to imitate a middle click event dispatch, I've
stripped out the javascript handling and just stuck a <div> element inside the
<a>. This causes the link to expand to the width of the cell.

Testing done:
Tested with various different dashboard configurations. All links that are done
via the cells (review request summaries, usernames) span the full width of the
cell, and work properly even when the text is wrapped. Other cells are
unaffected.

Reviewed at https://reviews.reviewboard.org/r/6011/

Fixes bug 2521.