Don't break when trying to check if TemplateHooks apply on 404s.
On 404 pages, the request doesn't have a resolver_match. For TemplateHooks
that apply to base.html (like the base-css hook point, for example), this
was problematic because 404.html extends base.html, and so any TemplateHook
that injects itself in a base hook point on only certain pages would cause
breakages since the resolver_match couldn't be checked to see if the
TemplateHook should actually apply.

Reviewed at https://reviews.reviewboard.org/r/6035/.