Allow resources to specify the name to use for links.
WebAPIResource subclasses can now set the link_name attribute to the
desired name for any links from other resources. Previously, this was
hard-coded to name_plural, and it still defaults to that.

This makes it easier to change a resource's name without affecting the
link, or to just choose a nicer name for links.

Reviewed at https://reviews.reviewboard.org/r/6125/