Fix 'ctrl+s' in inlineEditors that use CodeMirror.
inlineEditor has two functions that save the contents: `save` and `submit`. Of
these two, `save` is really just meant to be an internal helper, and `submit`
is the one that does a lot of the necessary cleanup.

The handler for 'ctrl+s' was calling `save`, which caused the editor box to
disappear. Changing it to `submit` makes everything work correctly.

Testing done:
Opened an inline editor that used CodeMirror and hit ctrl+s. Saw the input box
disappear and be replaced with the rendered contents.

Fixes bug 3489.

Reviewed at https://reviews.reviewboard.org/r/6133/