Remove the deprecation warning from djblet.util.fields.
The djblets.util.fields deprecation warning is annoying, and we'll never
be able to make it go away without just turning it off. The reason for
this is that Django Evolution signatures may reference the fields from
this file, and those are baked into history at that point. So really,
this file has to exist forever.

This change removes the warning, and instead adds a comment at the top
about this file being deprecated.

Reviewed at https://reviews.reviewboard.org/r/6147/