Allow greater control over inline editor buttons and rendering.
The buttons element is no longer always a <div>. If it's a multi-line
editor, it will appear as a <div/>, but if it's a single-line editor, it
will appear as a <span/>. This is preferred over hard-coding a
`display: inline` style on the element, and will have the same effect by
default. Custom CSS can then override it to have whatever display style
is appropriate.

There are also cases where a caller may need to set some properties on
an element before actually showing the editor and triggering the layout
code. For example, an element may need to be set that would impact the
positioning of the buttons or some other aspect of the editor.

To handle this, there's now a beginEditPreShow event that callers can
listen to. This is displayed immediately before the editor is shown.

Reviewed at https://reviews.reviewboard.org/r/6173/