Fix the sizing of inline editor fields on Firefox.
The inline editor fields were appearing very short on Firefox when set
to have buttons appear on the following line. This was due to our
calculation for whether buttons were on the next line, in combination
with however Firefox is representing coordinates and sizings.

Instead of trying to be fancy and detect things like the buttons
wrapping to the next line, we're now just checking the 'display'
property of the buttons. If it's 'inline' or 'inline-block', we'll
factor in the size. Otherwise, we won't. It's up to the caller to ensure
that the right 'display' is set and that elemens either wrap or don't.

Reviewed at https://reviews.reviewboard.org/r/6277/