Fix $.fn.retinaGravatar to not stomp all over parameters.
The implementation of URL parsing for retina gravatars was incredibly clumsy if
the user was using any of the extra parameters which are available in settings.
For example, if the user had set GRAVATAR_DEFAULT to something fun, retina
users wouldn't see it because it wouldn't preserve the d= param.

This change makes the parameter parsing much more capable, fixing the size
parameter without touching any of the others.

Testing done:
Set GRAVATAR_DEFAULT to 'wavatar' in my settings_local.py file and saw actual
wavatars instead of the ugly Gs.

Reviewed at https://reviews.reviewboard.org/r/6301/